+++ web-ui/src/main/java/org/archcnl/domain/input/model/presets/ArchitecturalStyle.java
@@ -1,47 +1,17 @@
 package org.archcnl.domain.input.model.presets;
 
-import com.fasterxml.jackson.annotation.JsonProperty;
-import java.util.ArrayList;
-import java.util.List;
+public enum ArchitecturalStyles {
+    MICROSERVICE_ARCHITECTURE("Microservice Architecture"),
+    LAYERED_ARCHITECTURE("Layered Architecture");
 
-public class ArchitecturalStyle {
+    private String architecturalStyleName;
 
-    @JsonProperty("style")
-    private ArchitecturalStyles style;
-
-    @JsonProperty("architectureInformation")
-    private List<ArchitectureInformation> variableParts = new ArrayList<ArchitectureInformation>();
-
-    @JsonProperty("architectureRules")
-    private List<ArchitectureRuleString> rules = new ArrayList<ArchitectureRuleString>();
-
-    @JsonProperty("style")
-    public ArchitecturalStyles getStyle() {
-        return style;
+    ArchitecturalStyles(String name) {
+        this.architecturalStyleName = name;
     }
 
-    @JsonProperty("style")
-    public void setStyle(ArchitecturalStyles name) {
-        this.style = name;
-    }
-
-    @JsonProperty("architectureInformation")
-    public List<ArchitectureInformation> getVariableParts() {
-        return variableParts;
-    }
-
-    @JsonProperty("architectureInformation")
-    public void setVariableParts(List<ArchitectureInformation> variableParts) {
-        this.variableParts = variableParts;
-    }
-
-    @JsonProperty("architectureRules")
-    public List<ArchitectureRuleString> getRules() {
-        return rules;
-    }
-
-    @JsonProperty("architectureRules")
-    public void setRules(List<ArchitectureRuleString> rules) {
-        this.rules = rules;
+    @Override
+    public String toString() {
+        return architecturalStyleName;
     }
 }
