--- web-ui/src/main/java/org/archcnl/domain/input/model/presets/ArchitecturalStyles.java
@@ -1,17 +1,0 @@
-package org.archcnl.domain.input.model.presets;
-
-public enum ArchitecturalStyles {
-    MICROSERVICE_ARCHITECTURE("Microservice Architecture"),
-    LAYERED_ARCHITECTURE("Layered Architecture");
-
-    private String architecturalStyleName;
-
-    ArchitecturalStyles(String name) {
-        this.architecturalStyleName = name;
-    }
-
-    @Override
-    public String toString() {
-        return architecturalStyleName;
-    }
-}
