+++ web-ui/src/main/java/org/archcnl/ui/inputview/presets/ArchitecturalStyleConfigManager.java
